refactor(de): split config and Deserializer into submodules
de.rs had grown to 2,658 lines spanning three concerns. Extract two of
them into a de/ submodule tree, leaving the entry-point functions
(from_str/from_slice/from_reader families + include/property resolution)
in de.rs:

- de/config.rs       — ParserConfig + builder, YamlVersion,
  RequireIndent, MergeKeyPolicy, DuplicateKeyPolicy (982 lines)
- de/deserializer.rs — the serde Deserializer<'de>, its access types,
  is_binary_tag, type_name (859 lines)

Pure mechanical move. Re-exported so every path resolves unchanged:
- public: ParserConfig, YamlVersion, RequireIndent, MergeKeyPolicy,
  DuplicateKeyPolicy, Deserializer
- crate-internal: SpannedMapAccess (used by value/serde_impl.rs) and
  is_binary_tag (used by streaming.rs) via crate::de::*.

de.rs drops 2,658 -> 839 lines.

Verified: build, fmt, clippy -D warnings, and `cargo test --all-features`
(527 lib + all integration suites + doctests) green.



---
THE ARCHITECT ᛫ Sebastien Rousseau ᛫ https://sebastienrousseau.com
THE ENGINE ᛞ EUXIS ᛫ Enterprise Unified Execution Intelligence System ᛫ https://euxis.co
